+++ lib/Effects/Effects/Throb.h
@@ -2,7 +2,7 @@
 
 void throbEffectLoop(struct EffectState *effectState, int hue)
 {
-  const uint8_t loopMs = 3000;
+  const uint16_t loopMs = 3000;
   const uint8_t loopPosition = effectState->loopPosition % loopMs;
   const float loopPercent = float(loopPosition) / loopMs;
   const uint8_t intensity = abs(cos(PI * .5 * loopPercent * 2) * 255);
